[Core] Make (number of) edges of line geometries complete and consistent with those of other geometries (#13916)
This PR aims at making the (number of) edges of line geometries complete and consistent with those of other geometries. When looking at other geometries (e.g. triangles or quadrilaterals), the number of edges corresponds to the number of sides of the geometry. For instance, for triangles that would always be 3, regardless of the order of the shape functions (and hence the number of intermediate nodes on a side). For line geometries, however, this was inconsistent. For instance, when inquiring the number of edges of a 3-point line geometry, the returned value would be 2. However, that is not the number of edges of the geometry, but the number of line segments that make up the side (or edge). For line geometries, to be consistent with other geometries, the returned value should always be 1.

In addition, several line geometries lacked an `override` of `GenerateEdges`, which yielded an error when attempting to inquire the edges of such a line geometry (raised by the base class, which said that an `override` had to be provided).

The modified code as well as the newly added code is covered by unit tests.

Changes include:
- Added `EdgeType` aliases to class templates `Line2D3`, `Line2D4`, `Line2D5`, and `Line3D3`, since these were missing.
- All line geometries now return 1 when inquiring the number of edges (i.e. when calling member `EdgesNumber`). This is in line with how other geometries define an edge.
- Added implementations of member `GenerateEdges` to class templates `Line2D3`, `Line2D4`, `Line2D5`, and `Line3D3`, since these were missing, too. In all cases, the returned geometry is identical to the line geometry itself.
- Corrected the implementations of `FacesNumber` of class templates `Line2D4` and `Line2D5` (i.e. when calling member `FacesNumber`). They now both return 0, since a line geometry doesn't have any faces. Note that this implementation is in line with those of the other line geometries.
- Based on the above listed changes, a few unit tests had to be corrected. Furthermore, new unit tests have been added to cover the new code.
- Avoid duplication of documentation by using `@copydoc`.
- Use the modern way of defining type aliases. That is, occurrences of `typedef` have been replaced by `using`.
- Removed some duplicated comments.